[Components]: Generalize the submit component and transforms it to button component.
