+++ resources/views/components/button.blade.php
@@ -1,0 +1,4 @@
+<button type="{{ $type }}" {{ $attributes->merge(['class' => "btn btn-{$theme}"]) }}>
+    @isset($icon) <i class="{{ $icon }}"></i> @endisset
+    {{ $label }}
+</button>
