--- resources/views/components/submit.blade.php
@@ -1,3 +1,0 @@
-<div class="{{$topclass}}">
-    <button type="submit" class="btn btn-{{$type}} {{$inputclass}}"><i class="{{$icon}}"></i> {{$label}}</button>
-</div>
