+++ tests/Components/ComponentsTest.php
@@ -29,6 +29,7 @@ class ComponentsTest extends TestCase
 
             // Form components.
 
+            "{$base}.button"       => new Components\Button(),
             "{$base}.date-range"   => new Components\DateRange('id'),
             "{$base}.input"        => new Components\Input('name'),
             "{$base}.input-color"  => new Components\InputColor('id'),
@@ -41,7 +42,6 @@ class ComponentsTest extends TestCase
             "{$base}.select"       => new Components\Select('name'),
             "{$base}.select2"      => new Components\Select2('id'),
             "{$base}.select-icon"  => new Components\SelectIcon('id'),
-            "{$base}.submit"       => new Components\Submit(),
             "{$base}.textarea"     => new Components\Textarea('name'),
             "{$base}.text-editor"  => new Components\TextEditor('id'),
 
@@ -106,6 +106,16 @@ class ComponentsTest extends TestCase
             $component = new Components\DateRange('id', null, null, null, $i);
             $this->assertIsString($component->initiator());
         }
+    }
+
+    public function testSelectComponent()
+    {
+        $component = new Components\Select('name');
+        $iClass = $component->makeItemClass(true);
+
+        $this->assertStringContainsString('form-control', $iClass);
+        $this->assertStringContainsString('w-100', $iClass);
+        $this->assertStringContainsString('is-invalid', $iClass);
     }
 
     public function testTextEditorComponent()
